+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/tabstray/TabsFeature.kt
@@ -14,17 +14,22 @@ import mozilla.components.support.base.feature.LifecycleAwareFeature
 
 /**
  * Feature implementation for connecting a tabs tray implementation with the session module.
+ *
+ * @param defaultTabsFilter A tab filter that is used for the initial presenting of tabs that will be used by
+ * [TabsFeature.filterTabs] by default as well.
  */
 class TabsFeature(
     tabsTray: TabsTray,
     private val store: BrowserStore,
     tabsUseCases: TabsUseCases,
+    private val defaultTabsFilter: (TabSessionState) -> Boolean = { true },
     closeTabsTray: () -> Unit
 ) : LifecycleAwareFeature {
     @VisibleForTesting
     internal var presenter = TabsTrayPresenter(
         tabsTray,
         store,
+        defaultTabsFilter,
         closeTabsTray
     )
 
@@ -49,9 +54,9 @@ class TabsFeature(
      * Filter the list of tabs using [tabsFilter].
      *
      * @param tabsFilter A filter function returning `true` for all tabs that should be displayed in
-     * the tabs tray.
+     * the tabs tray. Uses the [defaultTabsFilter] if none is provided.
      */
-    fun filterTabs(tabsFilter: (TabSessionState) -> Boolean) {
+    fun filterTabs(tabsFilter: (TabSessionState) -> Boolean = defaultTabsFilter) {
         presenter.tabsFilter = tabsFilter
 
         val state = store.state