+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/TabsTrayFragment.kt
@@ -47,7 +47,7 @@ class TabsTrayFragment : Fragment(), UserInteractionHandler {
             tabsTray,
             components.store,
             components.tabsUseCases,
-            ::closeTabsTray
+            closeTabsTray = ::closeTabsTray
         ).also { lifecycle.addObserver(it) }
     }